Introducing per-task/per-process cpufreq statistics.
The statistics are provided in /proc/<pid>/cpufreq_stats as a read-only
interface. This gives the amount of time spent by a task/process in
each of the frequencies on each of the CPUs. The cat output will have
"<cpu> <frequency> <time>" triplet in each line, which will mean that
a given task spent <time> usertime units of time at <frequency> on
CPU <cpu>. usertime units here is 10mS (similar to other time exported
in /proc). The statistics are available during the lifetime of the
tasks/processes.

Change-Id: I3ebb338c27020aabd6b4e258bd26981c51a57771

Conflicts:
	include/linux/sched.h